# doc/api.spec.yaml
openapi: 3.0.3
info:
  title: Tournament API
  description: Table tennis tournament API
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/api/v1.0.0'
    description: Tournament api (local)
  - url: 'http://dev.tournament.api/api/v1.0.0'
    description: Tournament api (dev)
components:
  parameters:
    playerId:
      name: playerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    groupId:
      name: groupId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    matchId:
      name: matchId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    offset:
      description: Pagination parameter for list operations
      name: offset
      in: query
      required: false
      schema:
        type: integer
    limit:
      description: Pagination parameter for list operations
      name: limit
      in: query
      required: false
      schema:
        type: integer
    sortBy:
      description: Order by field parameter for list operations
      name: sortBy
      in: query
      required: false
      schema:
        type: string
    sortOrder:
      description: Sort by ASC|DESC parameter for list operations
      name: sortOrder
      in: query
      required: false
      schema:
        type: string
        enum:
          - ASC
          - DESC
  schemas:
    CommonResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
...
    SeriesRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum:
            - draft
            - in_progress
            - done
...
    GroupRequest:
      type: object
      allOf:
        - $ref: "#/components/schemas/SeriesRequest"
      properties:
        tournamentId:
          type: integer
          format: int64
        runDate:
          type: string
          format: date-time
        position:
          type: integer
    GroupResponse:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - $ref: "#/components/schemas/GroupRequest"
    GroupPlayerResponse:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
      properties:
        groupId:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        startingPosition:
          type: integer
    MatchRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - DRAFT
            - IN_PROGRESS
            - DONE
        groupId:
          type: integer
          format: int64
        groupPlayerId1:
          type: integer
          format: int64
        groupPlayerId2:
          type: integer
          format: int64
        playerResult1:
          type: integer
        playerResult2:
          type: integer
    MatchResponse:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - $ref: "#/components/schemas/MatchRequest"
  examples:
...
    ExampleGroupResponse:
      description: Group created
      value:
        id: 1
        tournamentId: 1
        runDate: 2021-09-28T08:00:00.000Z
        position: 1
        status: DRAFT
        name: test group
        created: 2021-09-28T08:00:00.000Z
        updated: 2021-09-28T08:00:00.000Z
    ExampleMatchListResponse:
      description: List of all match
      value:
        - id: 1
          groupId: 1
          groupPlayerId1: 1
          groupPlayerId2: 2
          playerResult1: 0
          playerResult2: 0
          created: 2021-09-28T08:00:00.000Z
          updated: 2021-09-28T08:00:00.000Z
        - id: 2
          groupId: 1
          groupPlayerId1: 1
          groupPlayerId2: 2
          playerResult1: 0
          playerResult2: 0
          created: 2021-09-28T08:00:00.000Z
          updated: 2021-09-28T08:00:00.000Z
    ExampleMatchResponse:
      description: Match created
      value:
        id: 1
        groupId: 1
        groupPlayerId1: 1
        groupPlayerId2: 2
        playerResult1: 0
        playerResult2: 0
        status: DRAFT
        created: 2021-09-28T08:00:00.000Z
        updated: 2021-09-28T08:00:00.000Z
    ExampleGroupPlayerResponse:
      description: Group seeding created
      value:
        id: 1
        groupId: 1
        playerId: 1
        startingPosition: 1
        created: 2021-09-28T08:00:00.000Z
        updated: 2021-09-28T08:00:00.000Z
...
paths:
...
  /groups/{groupId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
    get:
      tags:
        - groups
      operationId: getGroup
      description: Get single series

      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/GroupRequest"
    put:
      tags:
        - groups
      operationId: updateGroup
      description: Update single series
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/GroupRequest"
      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/GroupResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleGroupResponse"
    delete:
      tags:
        - groups
      operationId: deleteGroup
      description: Remove series
      responses:
        '200':
          description: Ok response
        '404':
          description: Not Found
        '403':
          description: Forbidden

  ########################
  # match api
  ########################
  /groups/{groupId}/matches:
    get:
      tags:
        - groups
      operationId: listMatches
      parameters:
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortOrder"
      description: List all matches in group
      responses:
        '200':
          description: Successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleMatchListResponse"
  /matches/{matchId}:
    parameters:
      - $ref: "#/components/parameters/matchId"
    get:
      tags:
        - matches
      operationId: getMatch
      description: Get single match

      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/MatchRequest"
    put:
      tags:
        - matches
      operationId: updateMatch
      description: Update single match
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/MatchRequest"
      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/MatchResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleMatchResponse"

  ########################
  # seeding api
  ########################
  /groups/{groupId}/players:
    get:
      tags:
        - groups
      operationId: listGroupPlayers
      parameters:
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortOrder"
      description: List all player in group
      responses:
        '200':
          description: Successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupPlayerResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleGroupPlayerResponse"
  /groups/{groupId}/players/{playerId}/seedings:
    get:
      tags:
        - seeding
      operationId: getGroupPlayer
      parameters:
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/playerId"
      description: get player of group
      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/GroupPlayerResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleGroupPlayerResponse"
    post:
      tags:
        - seeding
      operationId: addPlayerToGroup
      parameters:
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/playerId"
      description: add player to group
      responses:
        '201':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/GroupPlayerResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleGroupPlayerResponse"
    delete:
      tags:
        - seeding
      operationId: deletePlayerFromGroup
      parameters:
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/playerId"
      description: Remove series
      responses:
        '200':
          description: Ok response
        '404':
          description: Not Found
        '403':
          description: Forbidden